pyproject.toml
[project]
name = "video-scene-thumbs"
version = "0.1.0"
description = "Extract scene thumbnails from video and create PDF"
requires-python = ">=3.8"
dependencies = [
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "scenedetect>=0.6.2",
    "streamlit>=1.32.0",
    "pyinstaller>=6.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"